Renamed screens as modules Added alias for modules Added screens and components for all modules Moved all ui components in the root of components deleted index.ts from default components Deleted bills and auth components and moved them in modules
