+++ src/app/components/NotFound.tsx
@@ -1,7 +1,8 @@
 import React from "react";
 import { ContainerProps } from "@d/theme/styles";
-import { Text, View } from "./defaults";
 import { NotFoundIcon } from "./Icons";
+import Text from "./defaults/Text";
+import View from "./defaults/View";
 
 
 interface NotFoundProps {
