+++ src/app/components/defaults/SafeAreaView.tsx
