--- src/app/components/ui/defaults/index.ts
@@ -1,12 +1,0 @@
-import Image from "./Image";
-import SafeAreaView from "./SafeAreaView";
-import Text from "./Text";
-import View from "./View";
-
-
-export {
-  Image,
-  SafeAreaView,
-  Text,
-  View,
-};
