+++ src/app/modules/auth/index.ts
@@ -1,0 +1,10 @@
+import SignInScreen from './screens/SignInScreen';
+import OTPVerifyScreen from './screens/OTPVerifyScreen';
+import ProfileScreen from './screens/ProfileScreen';
+
+
+export {
+  SignInScreen,
+  OTPVerifyScreen,
+  ProfileScreen,
+};
