+++ src/app/modules/auth/screens/OTPVerifyScreen.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState, useRef } from "react";
 import { View, Button, Text, TextInput } from "react-native";
-import { AuthContainer } from "@components/auth";
-import { Loader } from "@components/ui";
+import { AuthContainer } from "@mods/auth/components";
+import { Loader } from "@components/index";
 import { useAppDispatch, useAppSelector } from "@store/hooks";
 import { verifyCode } from "@store/slices/auth";
 import Styles from "./styles";
