+++ src/app/modules/auth/screens/SignInScreen.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState, useRef } from "react";
 import { View, Text, TextInput, Button } from "react-native";
 import { FirebaseRecaptchaVerifierModal, FirebaseAuthApplicationVerifier } from "expo-firebase-recaptcha";
-import { AuthContainer } from "@components/auth";
-import { Loader } from "@components/ui";
+import { AuthContainer } from "@mods/auth/components";
+import { Loader } from "@components/index";
 import Firebase from "@config/firebase";
 import { useAppDispatch, useAppSelector } from "@store/hooks";
 import { initPhoneSignIn } from "@store/slices/auth";
