+++ src/app/modules/bill/index.ts
@@ -1,0 +1,6 @@
+import BillManagerScreen from "./screens/BillManagerScreen";
+
+
+export {
+  BillManagerScreen,
+};
