+++ src/app/modules/bill/screens/BillManagerScreen.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 import { TouchableWithoutFeedback } from "react-native";
 import { useTheme } from "react-native-elements";
-import { ActionBar, FloatAddIcon, ListContainer, NotFound, SafeAreaView } from "@components/ui";
+import { ActionBar, FloatAddIcon, ListContainer, NotFound, SafeAreaView } from "@components/index";
 import { BillManagerScreenNavigationProps } from "@d/navigation";
-import { Bill } from "@components/bill";
+import { Bill } from "@mods/bill/components";
 
 
 interface BillManagerScreenProps extends BillManagerScreenNavigationProps { };
