--- src/app/screens/auth/index.ts
@@ -1,10 +1,0 @@
-import SignInScreen from './SignInScreen';
-import OTPVerifyScreen from './OTPVerifyScreen';
-import ProfileScreen from './ProfileScreen';
-
-
-export {
-  SignInScreen,
-  OTPVerifyScreen,
-  ProfileScreen,
-};
