--- src/app/screens/bill/index.ts
@@ -1,6 +1,0 @@
-import BillManagerScreen from "./BillManagerScreen";
-
-
-export {
-  BillManagerScreen,
-};
